mono calculation run by Bluesky
